Rename Init() to ensureNotifyShutdown.
This makes it private, removing it from the public api.

We don't really need clients to call init. We just need to make sure NotifyShutdown is called before we Wait or Close are called,
so we now call that inside WaitContext or Close.

I have prs for all the places that were using this to remove the init call.

I need this chagne as part of some stuff I'm doing in the api service.